+++ js/sidebar/subreddit_sidebar.js
